+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
 
+## 0.1.3 - Febuary 27, 2022
+
+Major updates:
+
+ - The code structure is modulized. The algorizmatic code is all in supervisor.py
+ - The focus command was roll backed and the user can also make the API and run the sequence
+ - linting, and also minor bugs were fixed.
+ - there was a error the the arc lamp was wrongly added on the fits header.
+
 ## 0.1.2 - January 17, 2022
 
 Major updates: